+++ pages/create-user.js
@@ -1,31 +1,8 @@
-import { useState } from 'react'
 import Layout from '../components/layout'
 import styles from '../styles/Home.module.css'
+import CreateForm from '../components/create-form'
 
 export default function CreateUser() {
-  const [userFirstName, setUserFirstName] = useState('');
-  const [userLastName, setUserLastName] = useState('');
-  const [isPending, setIsPending] = useState(false);
-
-  const handleSubmit = (e) => {
-    e.preventDefault();
-
-    const user = { first_name: userFirstName, last_name: userLastName };
-    setIsPending(true);
-
-    fetch('https://www.rdv-solidarites.fr/api/v1/users', {
-      method: 'POST',
-      headers: { 'Content-Type': 'application/json' },
-      body: JSON.stringify(user)
-    })
-
-      .then(() => {
-      setUserFirstName('');
-      setUserLastName('');
-      setIsPending(false);
-      alert(`Fiche utilisateur créée sur RDV Solidarités avec succès pour ${userFirstName} ${userLastName}`);
-    })
-  }
 
   return(
     <Layout className={styles.container}>
@@ -34,26 +11,7 @@ export default function CreateUser() {
           Créer une fiche bénéficiaire sur RDV Solidarités
         </h1>
         <div className={styles.create}>
-          <form
-            onSubmit={handleSubmit}
-          >
-            <label>Prénom</label>
-            <input
-              type="text"
-              value={userFirstName}
-              onChange={(e) => setUserFirstName(e.target.value)}
-              required
-            />
-            <label>Nom</label>
-            <input
-              type="text"
-              value={userLastName}
-              onChange={(e) => setUserLastName(e.target.value)}
-              required
-            />
-            {!isPending && <button>Créer fiche utilisateur</button>}
-            {isPending && <button>Fiche utilisateur en cours de création</button>}
-          </form>
+          <CreateForm></CreateForm>
         </div>
 
       </main>
